# Optimised CopyCPU and CopyCUDA for case when dst is contiguous and src has single element. [Fill]