--- a/cpp/benchmarks/CMakeLists.txt
+++ b/cpp/benchmarks/CMakeLists.txt
@@ -46,13 +46,14 @@ set(BENCHMARK_SOURCE_FILES
 )
 
 add_executable(benchmarks ${BENCHMARK_SOURCE_FILES})
-add_definitions(-DTEST_DATA_DIR="${PROJECT_SOURCE_DIR}/examples/test_data")
-add_definitions(-DBENCHMARK_DATA_DIR="${PROJECT_SOURCE_DIR}/data/Benchmark")
+target_compile_definitions(benchmarks PRIVATE TEST_DATA_DIR="${PROJECT_SOURCE_DIR}/examples/test_data")
+target_compile_definitions(benchmarks PRIVATE BENCHMARK_DATA_DIR="${PROJECT_SOURCE_DIR}/data/Benchmark")
 
-target_link_libraries(benchmarks ${CMAKE_PROJECT_NAME} benchmark::benchmark benchmark::benchmark_main)
+target_link_libraries(benchmarks PRIVATE ${CMAKE_PROJECT_NAME} benchmark::benchmark benchmark::benchmark_main)
 open3d_show_and_abort_on_warning(benchmarks)
 open3d_set_global_properties(benchmarks)
+open3d_link_3rdparty_libraries(benchmarks)
 
 if (BUILD_CUDA_MODULE)
-    target_include_directories(benchmarks SYSTEM PUBLIC ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES})
+    target_include_directories(benchmarks SYSTEM PRIVATE ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES})
 endif()