--- a/cpp/open3d/core/kernel/UnaryEWCPU.cpp
+++ b/cpp/open3d/core/kernel/UnaryEWCPU.cpp
@@ -132,6 +132,7 @@ void CopyCPU(const Tensor& src, Tensor& dst) {
                               src_dtype.ByteSize() * shape.NumElements());
     } else if (dst.IsContiguous() && src.NumElements() == 1 &&
                !src_dtype.IsObject()) {
+        int64_t num_elements = dst.NumElements();
         DISPATCH_DTYPE_TO_TEMPLATE_WITH_BOOL(src_dtype, [&]() {
             using src_t = scalar_t;
             DISPATCH_DTYPE_TO_TEMPLATE_WITH_BOOL(dst_dtype, [&]() {
@@ -139,7 +140,7 @@ void CopyCPU(const Tensor& src, Tensor& dst) {
                 const dst_t value = static_cast<dst_t>(src.Item<src_t>());
                 dst_t* dst_ptr = dst.GetDataPtr<dst_t>();
                 CPULauncher::LaunchGeneralKernel(
-                        dst.NumElements(), [&](int64_t workload_idx) {
+                        num_elements, [&](int64_t workload_idx) {
                             dst_ptr[workload_idx] = value;
                         });
             });