Benchmark/Zeros: 
Before: 
![image](https://user-images.githubusercontent.com/32535717/110248883-43d9c080-7f99-11eb-94f3-4f7592fa9490.png)

After:
```
-----------------------------------------------------
Benchmark           Time             CPU   Iterations
-----------------------------------------------------
Zeros/CPU         124 ms          124 ms            6
Zeros/CUDA       10.7 ms         10.7 ms           66

```
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/intel-isl/open3d/3136)
<!-- Reviewable:end -->